+++ Source/common/processtree/SNTEndpointSecurityAdapter.h
@@ -11,29 +11,20 @@
 /// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 /// See the License for the specific language governing permissions and
 /// limitations under the License.
-#ifndef SANTA__SANTAD_PROCESSTREE_ANNOTATIONS_BASE_H
-#define SANTA__SANTAD_PROCESSTREE_ANNOTATIONS_BASE_H
+#ifndef SANTA__COMMON__PROCESSTREE__SNTENDPOINTSECURITYADAPTER_H
+#define SANTA__COMMON__PROCESSTREE__SNTENDPOINTSECURITYADAPTER_H
 
-#include <optional>
+#include <EndpointSecurity/EndpointSecurity.h>
 
-#include "Source/santad/ProcessTree/process_tree.pb.h"
+#include "Source/common/es/Message.h"
+#include "Source/common/processtree/process_tree.h"
 
 namespace santa::santad::process_tree {
 
-class ProcessTree;
-class Process;
-
-class Annotator {
- public:
-  virtual ~Annotator() = default;
-
-  virtual void AnnotateFork(ProcessTree &tree, const Process &parent,
-                            const Process &child) = 0;
-  virtual void AnnotateExec(ProcessTree &tree, const Process &orig_process,
-                            const Process &new_process) = 0;
-  virtual std::optional<::santa::pb::v1::process_tree::Annotations> Proto()
-      const = 0;
-};
+// Inform the tree of the ES event in msg.
+// This is idempotent on the tree, so can be called from multiple places with
+// the same msg.
+void InformFromESEvent(ProcessTree &tree, const santa::Message &msg);
 
 }  // namespace santa::santad::process_tree
