+++ Source/common/processtree/annotations/BUILD
@@ -2,14 +2,14 @@ load("@rules_cc//cc:defs.bzl", "cc_library")
 load("//:helper.bzl", "santa_unit_test")
 
 package(
-    default_visibility = ["//:santa_package_group"],
+    default_visibility = ["//visibility:public"],
 )
 
 cc_library(
     name = "annotator",
     hdrs = ["annotator.h"],
     deps = [
-        "//Source/santad/ProcessTree:process_tree_cc_proto",
+        "//Source/common/processtree:process_tree_cc_proto",
     ],
 )
 
@@ -19,9 +19,9 @@ cc_library(
     hdrs = ["originator.h"],
     deps = [
         ":annotator",
-        "//Source/santad/ProcessTree:process",
-        "//Source/santad/ProcessTree:process_tree",
-        "//Source/santad/ProcessTree:process_tree_cc_proto",
+        "//Source/common/processtree:process",
+        "//Source/common/processtree:process_tree",
+        "//Source/common/processtree:process_tree_cc_proto",
         "@abseil-cpp//absl/container:flat_hash_map",
     ],
 )
@@ -31,8 +31,8 @@ santa_unit_test(
     srcs = ["originator_test.mm"],
     deps = [
         ":originator",
-        "//Source/santad/ProcessTree:process",
-        "//Source/santad/ProcessTree:process_tree_cc_proto",
-        "//Source/santad/ProcessTree:process_tree_test_helpers",
+        "//Source/common/processtree:process",
+        "//Source/common/processtree:process_tree_cc_proto",
+        "//Source/common/processtree:process_tree_test_helpers",
     ],
 )
