+++ Source/common/processtree/annotations/originator.cc
@@ -11,16 +11,16 @@
 /// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 /// See the License for the specific language governing permissions and
 /// limitations under the License.
-#include "Source/santad/ProcessTree/annotations/originator.h"
+#include "Source/common/processtree/annotations/originator.h"
 
 #include <memory>
 #include <optional>
 #include <string>
 #include <utility>
 
-#include "Source/santad/ProcessTree/process.h"
-#include "Source/santad/ProcessTree/process_tree.h"
-#include "Source/santad/ProcessTree/process_tree.pb.h"
+#include "Source/common/processtree/process.h"
+#include "Source/common/processtree/process_tree.h"
+#include "Source/common/processtree/process_tree.pb.h"
 #include "absl/container/flat_hash_map.h"
 
 namespace ptpb = ::santa::pb::v1::process_tree;
