+++ Source/common/processtree/process_tree.cc
@@ -11,7 +11,7 @@
 /// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 /// See the License for the specific language governing permissions and
 /// limitations under the License.
-#include "Source/santad/ProcessTree/process_tree.h"
+#include "Source/common/processtree/process_tree.h"
 
 #include <sys/types.h>
 
@@ -25,9 +25,9 @@
 #include <utility>
 #include <vector>
 
-#include "Source/santad/ProcessTree/annotations/annotator.h"
-#include "Source/santad/ProcessTree/process.h"
-#include "Source/santad/ProcessTree/process_tree.pb.h"
+#include "Source/common/processtree/annotations/annotator.h"
+#include "Source/common/processtree/process.h"
+#include "Source/common/processtree/process_tree.pb.h"
 #include "absl/container/flat_hash_map.h"
 #include "absl/container/flat_hash_set.h"
 #include "absl/status/status.h"
