+++ Source/common/processtree/process_tree_macos.h
@@ -11,8 +11,8 @@
 /// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 /// See the License for the specific language governing permissions and
 /// limitations under the License.
-#ifndef SANTA__SANTAD_PROCESSTREE_TREE_MACOS_H
-#define SANTA__SANTAD_PROCESSTREE_TREE_MACOS_H
+#ifndef SANTA__COMMON__PROCESSTREE__PROCESS_TREE_MACOS_H
+#define SANTA__COMMON__PROCESSTREE__PROCESS_TREE_MACOS_H
 
 #include <bsm/libbsm.h>
