+++ Source/common/processtree/process_tree_macos.mm
@@ -11,7 +11,7 @@
 /// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 /// See the License for the specific language governing permissions and
 /// limitations under the License.
-#include "Source/santad/ProcessTree/process_tree.h"
+#include "Source/common/processtree/process_tree.h"
 
 #import <Foundation/Foundation.h>
 #include <bsm/libbsm.h>
@@ -24,9 +24,9 @@
 #include <optional>
 #include <vector>
 
+#include "Source/common/CSOpsHelper.h"
 #include "Source/common/SystemResources.h"
-#include "Source/santad/CSOpsHelper.h"
-#include "Source/santad/ProcessTree/process.h"
+#include "Source/common/processtree/process.h"
 #include "absl/container/flat_hash_map.h"
 #include "absl/status/status.h"
 #include "absl/status/statusor.h"
