+++ Source/common/santa.proto
@@ -16,7 +16,7 @@ syntax = "proto3";
 
 package santa.pb.v1;
 
-import "Source/santad/ProcessTree/process_tree.proto";
+import "Source/common/processtree/process_tree.proto";
 import "google/protobuf/any.proto";
 import "google/protobuf/timestamp.proto";
