+++ Source/santad/EventProviders/SNTEndpointSecurityTreeAwareClient.h
@@ -16,7 +16,7 @@
 
 #include "Source/common/es/ESMetricsObserver.h"
 #include "Source/common/es/EndpointSecurityAPI.h"
-#include "Source/santad/ProcessTree/process_tree.h"
+#include "Source/common/processtree/process_tree.h"
 
 @interface SNTEndpointSecurityTreeAwareClient : SNTEndpointSecurityClient
 @property std::shared_ptr<santa::santad::process_tree::ProcessTree> processTree;
