+++ Source/santad/Santad.h
@@ -22,10 +22,10 @@
 #include "Source/common/es/EndpointSecurityAPI.h"
 #include "Source/common/es/Enricher.h"
 #include "Source/common/faa/WatchItems.h"
+#include "Source/common/processtree/process_tree.h"
 #include "Source/santad/EventProviders/AuthResultCache.h"
 #include "Source/santad/Logs/EndpointSecurity/Logger.h"
 #include "Source/santad/Metrics.h"
-#include "Source/santad/ProcessTree/process_tree.h"
 #import "Source/santad/SNTCompilerController.h"
 #import "Source/santad/SNTExecutionController.h"
 #import "Source/santad/SNTNetworkExtensionQueue.h"
