+++ Source/santad/SantadDeps.mm
@@ -29,11 +29,11 @@
 #include "Source/common/es/EnrichedTypes.h"
 #include "Source/common/es/Message.h"
 #include "Source/common/faa/WatchItems.h"
+#include "Source/common/processtree/annotations/originator.h"
+#include "Source/common/processtree/process_tree.h"
 #import "Source/santad/DataLayer/SNTEventTable.h"
 #import "Source/santad/DataLayer/SNTRuleTable.h"
 #include "Source/santad/EntitlementsFilter.h"
-#include "Source/santad/ProcessTree/annotations/originator.h"
-#include "Source/santad/ProcessTree/process_tree.h"
 #import "Source/santad/SNTDatabaseController.h"
 #include "Source/santad/SNTDecisionCache.h"
 #import "Source/santad/SNTNetworkExtensionQueue.h"
